Move `Formatter.columns` into `utils/formatter`.
